requirements.txt
requests
zipstream

tagstore
zoterobib

django-cms
djangocms-file
djangocms-link
djangocms-snippet
djangocms-text-ckeditor

git+https://github.com/DocOtak/djangocms-picture.git
